Passando responsabilidade de montar os cards para o componente Dashboard, ao invés da Página Dashboard
